refactor(commit-preview): extract commit preview question wiring into helper
